apps/sync/requirements.txt
dbt-postgres
psycopg2
requests
python-dotenv